Added header to index, changed command names
The command names have changed to denote the underlying algorithms:
index    => pseudoindex
map      => pseudomap
saindex  => quasiindex
samap    => quasimap

A header is now written along with each index that contains the index
type, version number, and other useful information.  This should help
in the future with detecting incompatible versions or when the wrong map
command is issued with this index.