Numerous fixes to SAM output
previously there were some bugs and corner cases in the SAM output.
Though the fields we were checking (transcript, positon etc.) were
output fine, there were issues with some other fields (e.g. CIGAR string)
that samtools didn't like.  This has been fixed now, so that it is possible
to pipe the output of rapmap directly to samtools and convert the output
to BAM.  *Note*: If a read is paired but the mate is un-mapped, Samtools
seems to want PNEXT of both the mapped read and the unmapped mate to be
equal to the POS of the mapped mate.