Added the *much* more memory frugal perfect hash
The FrugalBooMap is much more stingy on memory because:
1) it doesn't store the k-mer keys directly (inferring them from the
   text and suffix array instead).
2) it uses a small integer (uint8_t) to store the length of each
   suffix array interval, keeping an overflow table in case
   this is not large enough.

Overall, this reduces the memory taken by the "hash" from
24-bytes per entry (8+8+8) to ~9-bytes per entry (8+1).  This
is approximate b/c it doesn't include the (usually very small)
overflow table.

The downsides of this map are that:
1) Construction takes considerably longer
2) Lookup becomes slower

With the previous replacement of google's dense_hash_map with
sparsepp, the memory consumption between the "regular" index
(i.e. sparsepp) and the non-frugal perfect hash became rather
small.  So, the question on the table now is, do we use the
frugal perfect hash or the existing perfect hash.  It seems
that there is little space in the middle for the non-frugal
perfect hash.  If you have enough memory for that, you likely
have enough for sparsepp --- which is a bit faster.  However,
if you don't have enough for sparsepp, then you might need
something as aggressive as the frugal perfect hash anyway.