maxSlack => consensusFraction and empty hit rule
change fixed maxSlack to be computed from a consensus fraction.
Rather than a mapping being allowed to miss up to maxSlack SA intervals,
we define the consensus fraction as the fraction of intervals that the
mapping must contain (strict mapping has a consensus fraction of 1).
Then, the maxSlack can be computed from the consensus fraction.

added a rule that if the consensus fraction is < 1, and there are
no targets that obtain the required number of hits, then simply consider
chaining for all targets.