Address issue #46.
In the SAM output of RapMap, there was an implicit assumption that
if a read is unmapped, then the read is written down in the forward
orientation.  That is, if the read unmapped flag is set (0x4), then
the flag 0x10 was assumed to have an invalid (random) state ---
the unmapped read is always recorded in its forward orientation.
However, while the SAM spec does dictate that many of the flags can
be ignored (i.e. are not guaranteed to be valid) if 0x4 is set, it
actually doesn't seem to make this claim about 0x10.  These changes
should assure that, for an orphaned pair, the mateFwd flag will be
set.